+++ resources/views/home.blade.php
@@ -3,7 +3,7 @@
 @section('content')
 <div class="container">
     <div class="row justify-content-center">
-        <div class="col-md-8">
+        <div class="col-12">
             <div class="card">
                 <div class="card-header">Profile</div>
 
@@ -19,10 +19,9 @@
                     <a class="btn btn-primary" href="{{ route('bible') }}">
                         Read Bible
                     </a>
-                        <?php
-                            //  print_r($user);
-                        ?>
-                    
+ 
+                    <saved-highlights-component :highlights="{{ $highlights }}"></saved-highlights-component>
+                   
                 </div>
             </div>
         </div>